requirements.txt
future==0.18.2
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.2
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
torch==1.5.0
torchvision==0.6.0
